requirements.txt
Flask==2.3.3
pandas==2.1.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
